+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.10.0] - 2019-06-05
+### Added
+- exitGame function to handle quitting ame
+- score variable to hold game score
+
 ## [0.9.0] - 2019-06-01
 ### Changed
 - Variable names for used_look and current_room to be consistent with rest of project
@@ -83,7 +88,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - .gitignore
 - Changelog and Readme files
 
-[Unreleased]: https://github.com/afs2015/PepperRPG/compare/v0.9.0....HEAD
+[Unreleased]: https://github.com/afs2015/PepperRPG/compare/v0.10.0....HEAD
+[0.10.0]: https://github.com/afs2015/PepperRPG/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/afs2015/PepperRPG/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/afs2015/PepperRPG/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/afs2015/PepperRPG/compare/v0.6.1...v0.7.0